requirements.txt
-i https://pypi.org/simple
streamlit
faiss-cpu
langchain==0.3.26
langchain-core==0.3.68
langchain-community==0.3.27
langchain-groq==0.3.5
langsmith
python-dotenv
pydantic
sentence-transformers
